LeAudio: Improve function naming
This change improves some of the function naming
so that it was more obvious what each of them do
at the point where they are called, without looking
into the implementation. This improves code readability.
It also for consisency replaces a static DebugDump
function call that was called on an instance, while
all other calls around are called directly from the
class namespace.

Bug: 256967923
Test: atest --host bluetooth_le_audio_test bluetooth_le_audio_client_test --no-bazel-mode
Change-Id: Id51c18211ee3936b8890dc21c0ed139969b51810